Move singleton class shared.Building into its own file. NFC.
A singleton class with only static methods is no different
to python module.  This also means less code that all users
of `shared` needs to import.

This mean moving move than half the code in shared.py which
makes it a lot more manageable.
