+++ embuilder.py
@@ -20,6 +20,7 @@ import os
 import subprocess
 import sys
 
+from tools import building
 from tools import shared
 from tools import system_libs
 
@@ -125,7 +126,7 @@ def build(src, result_libs, args=[]):
       shared.Cache.erase_file(lib)
 
   try:
-    shared.Building.emcc(cpp, args, output_filename=temp_js)
+    building.emcc(cpp, args, output_filename=temp_js)
   except subprocess.CalledProcessError as e:
     shared.exit_with_error("embuilder: emcc command failed with %d: '%s'", e.returncode, ' '.join(e.cmd))
