+++ emcc.py
@@ -40,7 +40,8 @@ import base64
 from subprocess import PIPE
 
 import emscripten
-from tools import shared, system_libs, client_mods, js_optimizer, jsrun, colored_logger, diagnostics
+from tools import shared, system_libs, client_mods, js_optimizer, jsrun
+from tools import colored_logger, diagnostics, building
 from tools.shared import unsuffixed, unsuffixed_basename, WINDOWS, safe_move, run_process, asbytes, read_and_preprocess, exit_with_error, DEBUG
 from tools.response_file import substitute_response_files
 from tools.minimal_runtime_shell import generate_minimal_runtime_html
@@ -177,14 +178,14 @@ def save_intermediate(name, suffix='js'):
   if isinstance(final, list):
     logger.debug('(not saving intermediate %s because deferring linking)' % name)
     return
-  shared.Building.save_intermediate(final, name + '.' + suffix)
+  building.save_intermediate(final, name + '.' + suffix)
 
 
 def save_intermediate_with_wasm(name, wasm_binary):
   if not DEBUG:
     return
   save_intermediate(name) # save the js
-  shared.Building.save_intermediate(wasm_binary, name + '.wasm')
+  building.save_intermediate(wasm_binary, name + '.wasm')
 
 
 class TimeLogger(object):
@@ -353,11 +354,11 @@ class JSOptimizer(object):
     if self.cleanup_shell and 'last' in passes:
       passes += ['cleanup']
     logger.debug('applying js optimization passes: %s', ' '.join(passes))
-    final = shared.Building.js_optimizer(final, passes, use_source_map(self),
-                                         self.extra_info, just_split=just_split,
-                                         just_concat=just_concat,
-                                         output_filename=self.in_temp(os.path.basename(final) + '.jsopted.js'),
-                                         extra_closure_args=self.closure_args)
+    final = building.js_optimizer(final, passes, use_source_map(self),
+                                  self.extra_info, just_split=just_split,
+                                  just_concat=just_concat,
+                                  output_filename=self.in_temp(os.path.basename(final) + '.jsopted.js'),
+                                  extra_closure_args=self.closure_args)
     self.js_transform_tempfiles.append(final)
     save_intermediate(title, suffix='js' if 'emitJSON' not in passes else 'json')
 
@@ -611,7 +612,7 @@ def backend_binaryen_passes():
   if not shared.Settings.EXIT_RUNTIME:
     passes += ['--no-exit-runtime']
   if shared.Settings.OPT_LEVEL > 0 or shared.Settings.SHRINK_LEVEL > 0:
-    passes += [shared.Building.opt_level_to_str(shared.Settings.OPT_LEVEL, shared.Settings.SHRINK_LEVEL)]
+    passes += [building.opt_level_to_str(shared.Settings.OPT_LEVEL, shared.Settings.SHRINK_LEVEL)]
   elif shared.Settings.STANDALONE_WASM:
     # even if not optimizing, make an effort to remove all unused imports and
     # exports, to make the wasm as standalone as possible
@@ -800,7 +801,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       lines = [x for x in proc.stderr.splitlines() if shared.CLANG_CC in x and input_file in x]
       parts = shlex.split(lines[0].replace('\\', '\\\\'))
       parts = [x for x in parts if x not in ['-c', '-o', '-v', '-emit-llvm'] and input_file not in x and temp_target not in x]
-      print(' '.join(shared.Building.doublequote_spaces(parts[1:])))
+      print(' '.join(building.doublequote_spaces(parts[1:])))
     return 0
 
   def get_language_mode(args):
@@ -1105,16 +1106,16 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           exit_with_error('%s: No such file or directory ("%s" was expected to be an input file, based on the commandline arguments provided)', arg, arg)
 
         file_suffix = get_file_suffix(arg)
-        if file_suffix in SOURCE_ENDINGS + OBJECT_FILE_ENDINGS + DYNAMICLIB_ENDINGS + ASSEMBLY_ENDINGS + HEADER_ENDINGS or shared.Building.is_ar(arg): # we already removed -o <target>, so all these should be inputs
+        if file_suffix in SOURCE_ENDINGS + OBJECT_FILE_ENDINGS + DYNAMICLIB_ENDINGS + ASSEMBLY_ENDINGS + HEADER_ENDINGS or building.is_ar(arg): # we already removed -o <target>, so all these should be inputs
           newargs[i] = ''
           if file_suffix in SOURCE_ENDINGS or (has_dash_c and file_suffix == '.bc'):
             input_files.append((i, arg))
           elif file_suffix in HEADER_ENDINGS:
             input_files.append((i, arg))
             has_header_inputs = True
-          elif file_suffix in ASSEMBLY_ENDINGS or shared.Building.is_bitcode(arg) or shared.Building.is_ar(arg):
+          elif file_suffix in ASSEMBLY_ENDINGS or building.is_bitcode(arg) or building.is_ar(arg):
             input_files.append((i, arg))
-          elif shared.Building.is_wasm(arg):
+          elif building.is_wasm(arg):
             if not shared.Settings.WASM_BACKEND:
               exit_with_error('fastcomp is not compatible with wasm object files:' + arg)
             input_files.append((i, arg))
@@ -1132,8 +1133,8 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           else:
             diagnostics.warning('invalid-input', arg + ' is not a valid input file')
         elif file_suffix in STATICLIB_ENDINGS:
-          if not shared.Building.is_ar(arg):
-            if shared.Building.is_bitcode(arg):
+          if not building.is_ar(arg):
+            if building.is_bitcode(arg):
               message = arg + ': File has a suffix of a static library ' + str(STATICLIB_ENDINGS) + ', but instead is an LLVM bitcode file! When linking LLVM bitcode files, use one of the suffixes ' + str(OBJECT_FILE_ENDINGS)
             else:
               message = arg + ': Unknown format, not a static library!'
@@ -1297,7 +1298,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       exit_with_error('no input files\nnote that input files without a known suffix are ignored, make sure your input files end with one of: ' + str(SOURCE_ENDINGS + OBJECT_FILE_ENDINGS + DYNAMICLIB_ENDINGS + STATICLIB_ENDINGS + ASSEMBLY_ENDINGS + HEADER_ENDINGS))
 
     # Note the exports the user requested
-    shared.Building.user_requested_exports = shared.Settings.EXPORTED_FUNCTIONS[:]
+    building.user_requested_exports = shared.Settings.EXPORTED_FUNCTIONS[:]
 
     # -s ASSERTIONS=1 implies the heaviest stack overflow check mode. Set the implication here explicitly to avoid having to
     # do preprocessor "#if defined(ASSERTIONS) || defined(STACK_OVERFLOW_CHECK)" in .js files, which is not supported.
@@ -1690,10 +1691,10 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       # dynCall_ii is used to call pthread entry points in worker.js (as
       # metadce does not consider worker.js, which is external, we must
       # consider it a user export, i.e., one which can never be removed).
-      shared.Building.user_requested_exports += ['dynCall_ii']
+      building.user_requested_exports += ['dynCall_ii']
 
       if shared.Settings.MINIMAL_RUNTIME:
-        shared.Building.user_requested_exports += ['exit']
+        building.user_requested_exports += ['exit']
 
       if shared.Settings.PROXY_TO_PTHREAD:
         shared.Settings.EXPORTED_FUNCTIONS += ['_proxy_main']
@@ -1741,7 +1742,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       shared.Settings.MINIFY_ASMJS_IMPORT_NAMES = 1
 
     if shared.Settings.WASM:
-      if not shared.Building.need_asm_js_file():
+      if not building.need_asm_js_file():
         asm_target = asm_target.replace('.asm.js', '.temp.asm.js')
         misc_temp_files.note(asm_target)
 
@@ -1977,7 +1978,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
 
       # run safe-heap as a binaryen pass in fastcomp wasm, while in the wasm backend we
       # run it in binaryen_passes so that it can be synchronized with the sbrk ptr
-      if shared.Settings.SAFE_HEAP and shared.Building.is_wasm_only() and not shared.Settings.WASM_BACKEND:
+      if shared.Settings.SAFE_HEAP and building.is_wasm_only() and not shared.Settings.WASM_BACKEND:
         options.binaryen_passes += ['--safe-heap']
       if shared.Settings.EMULATE_FUNCTION_POINTER_CASTS and not shared.Settings.WASM_BACKEND:
         # emulated function pointer casts is emulated in fastcomp wasm using a binaryen pass
@@ -2022,7 +2023,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         exit_with_error('NODE_CODE_CACHING saves a file on the side and is not compatible with SINGLE_FILE')
 
     # safe heap in asm.js uses the js optimizer (in wasm-only mode we can use binaryen)
-    if shared.Settings.SAFE_HEAP and not shared.Building.is_wasm_only():
+    if shared.Settings.SAFE_HEAP and not building.is_wasm_only():
       if not options.js_opts:
         logger.debug('enabling js opts for SAFE_HEAP')
         options.js_opts = True
@@ -2055,7 +2056,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         newargs.append('-disable-llvm-optzns')
 
     if not shared.Settings.LEGALIZE_JS_FFI:
-      assert shared.Building.is_wasm_only(), 'LEGALIZE_JS_FFI incompatible with RUNNING_JS_OPTS.'
+      assert building.is_wasm_only(), 'LEGALIZE_JS_FFI incompatible with RUNNING_JS_OPTS.'
 
     # check if we can address the 2GB mark and higher: either if we start at
     # 2GB, or if we allow growth to either any amount or to 2GB or more.
@@ -2112,7 +2113,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       if shared.Settings.DEBUG_LEVEL >= 3 and not shared.Settings.WASM and not shared.Settings.WASM_BACKEND:
         cflags.append('-fno-discard-value-names')
 
-      if not shared.Building.can_inline():
+      if not building.can_inline():
         cflags.append('-fno-inline-functions')
 
       # For fastcomp backend, no LLVM IR functions should ever be annotated
@@ -2183,7 +2184,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           else:
             # With fastcomp (or with LTO mode) these args get passed instead
             # at link time when the backend runs.
-            for a in shared.Building.llvm_backend_args():
+            for a in building.llvm_backend_args():
               cmd += ['-mllvm', a]
         else:
           cmd.append('-emit-llvm')
@@ -2203,7 +2204,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         elif file_suffix in DYNAMICLIB_ENDINGS:
           logger.debug('using shared library: ' + input_file)
           temp_files.append((i, input_file))
-        elif shared.Building.is_ar(input_file):
+        elif building.is_ar(input_file):
           logger.debug('using static library: ' + input_file)
           ensure_archive_index(input_file)
           temp_files.append((i, input_file))
@@ -2211,7 +2212,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           temp_file = in_temp(unsuffixed(uniquename(input_file)) + '.o')
           if file_suffix == '.ll':
             logger.debug('assembling assembly file: ' + input_file)
-            shared.Building.llvm_as(input_file, temp_file)
+            building.llvm_as(input_file, temp_file)
             temp_files.append((i, temp_file))
           else:
             if not shared.Settings.WASM_BACKEND:
@@ -2243,7 +2244,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
               # will find them (otherwise, llvm.cos.f32() will not link in cosf(), and
               # we end up calling out to JS for Math.cos).
               opts = options.llvm_opts + ['-lower-non-em-intrinsics']
-              shared.Building.llvm_opt(temp_file, opts, new_temp_file)
+              building.llvm_opt(temp_file, opts, new_temp_file)
               temp_files[pos] = (temp_files[pos][0], new_temp_file)
 
       # If we were just compiling stop here
@@ -2310,7 +2311,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           # skip running the linker and just copy the object file
           shutil.copyfile(temp_file, specified_target)
         else:
-          shared.Building.link_to_object(linker_inputs, specified_target)
+          building.link_to_object(linker_inputs, specified_target)
         logger.debug('stopping after linking to object file')
         return 0
 
@@ -2359,7 +2360,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       if not perform_link:
         is_bc = suffix(temp_files[0][1]) in OBJECT_FILE_ENDINGS
         is_dylib = suffix(temp_files[0][1]) in DYNAMICLIB_ENDINGS
-        is_ar = shared.Building.is_ar(temp_files[0][1])
+        is_ar = building.is_ar(temp_files[0][1])
         perform_link = not (is_bc or is_dylib) and is_ar
       if perform_link:
         logger.debug('linking: ' + str(linker_inputs))
@@ -2376,13 +2377,13 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           if shared.Settings.LLD_REPORT_UNDEFINED and shared.Settings.ERROR_ON_UNDEFINED_SYMBOLS:
             js_funcs = get_all_js_syms(misc_temp_files)
             log_time('JS symbol generation')
-          final = shared.Building.link_lld(linker_inputs, DEFAULT_FINAL, external_symbol_list=js_funcs)
+          final = building.link_lld(linker_inputs, DEFAULT_FINAL, external_symbol_list=js_funcs)
           # Special handling for when the user passed '-Wl,--version'.  In this case the linker
           # does not create the output file, but just prints its version and exits with 0.
           if '--version' in linker_inputs:
             return 0
         else:
-          final = shared.Building.link(linker_inputs, DEFAULT_FINAL, force_archive_contents=force_archive_contents, just_calculate=just_calculate)
+          final = building.link(linker_inputs, DEFAULT_FINAL, force_archive_contents=force_archive_contents, just_calculate=just_calculate)
       else:
         logger.debug('skipping linking: ' + str(linker_inputs))
         temp_file = temp_files[0][1]
@@ -2417,23 +2418,23 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
 
         if options.llvm_lto is not None and options.llvm_lto >= 2 and optimizing(options.llvm_opts):
           logger.debug('running LLVM opts as pre-LTO')
-          final = shared.Building.llvm_opt(final, options.llvm_opts, DEFAULT_FINAL)
+          final = building.llvm_opt(final, options.llvm_opts, DEFAULT_FINAL)
           save_intermediate('opt', 'bc')
 
         # If we can LTO, do it before dce, since it opens up dce opportunities
         if (not shared.Settings.LINKABLE) and options.llvm_lto and options.llvm_lto != 2:
-          if not shared.Building.can_inline():
+          if not building.can_inline():
             link_opts.append('-disable-inlining')
           # add a manual internalize with the proper things we need to be kept alive during lto
-          link_opts += shared.Building.get_safe_internalize() + ['-std-link-opts']
+          link_opts += building.get_safe_internalize() + ['-std-link-opts']
           # execute it now, so it is done entirely before we get to the stage of legalization etc.
-          final = shared.Building.llvm_opt(final, link_opts, DEFAULT_FINAL)
+          final = building.llvm_opt(final, link_opts, DEFAULT_FINAL)
           save_intermediate('lto', 'bc')
           link_opts = []
         else:
           # At minimum remove dead functions etc., this potentially saves a
           # lot in the size of the generated code (and the time to compile it)
-          link_opts += shared.Building.get_safe_internalize() + ['-globaldce']
+          link_opts += building.get_safe_internalize() + ['-globaldce']
 
         if options.cfi:
           if use_cxx:
@@ -2443,11 +2444,11 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         if shared.Settings.AUTODEBUG:
           # let llvm opt directly emit ll, to skip writing and reading all the bitcode
           link_opts += ['-S']
-          final = shared.Building.llvm_opt(final, link_opts, get_final() + '.link.ll')
+          final = building.llvm_opt(final, link_opts, get_final() + '.link.ll')
           save_intermediate('linktime', 'll')
         else:
           if len(link_opts) > 0:
-            final = shared.Building.llvm_opt(final, link_opts, DEFAULT_FINAL)
+            final = building.llvm_opt(final, link_opts, DEFAULT_FINAL)
             save_intermediate('linktime', 'bc')
           if options.save_bc:
             shutil.copyfile(final, options.save_bc)
@@ -2553,7 +2554,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         final += '.tr.js'
         posix = not shared.WINDOWS
         logger.debug('applying transform: %s', options.js_transform)
-        shared.check_call(shared.Building.remove_quotes(shlex.split(options.js_transform, posix=posix) + [os.path.abspath(final)]))
+        shared.check_call(building.remove_quotes(shlex.split(options.js_transform, posix=posix) + [os.path.abspath(final)]))
         save_intermediate('transformed')
 
       js_transform_tempfiles = [final]
@@ -2630,7 +2631,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
 
         # Minify the worker.js file in optimized builds
         if (shared.Settings.OPT_LEVEL >= 1 or shared.Settings.SHRINK_LEVEL >= 1) and not shared.Settings.DEBUG_LEVEL:
-          minified_worker = shared.Building.acorn_optimizer(worker_output, ['minifyWhitespace'], return_output=True)
+          minified_worker = building.acorn_optimizer(worker_output, ['minifyWhitespace'], return_output=True)
           open(worker_output, 'w').write(minified_worker)
 
       # Generate the fetch.js worker script for multithreaded emscripten_fetch() support if targeting pthreads.
@@ -2687,7 +2688,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
             optimizer.queue += ['optimizeFrounds']
 
       if options.js_opts:
-        if shared.Settings.SAFE_HEAP and not shared.Building.is_wasm_only():
+        if shared.Settings.SAFE_HEAP and not building.is_wasm_only():
           optimizer.queue += ['safeHeap']
 
         if shared.Settings.OPT_LEVEL >= 2 and shared.Settings.DEBUG_LEVEL < 3:
@@ -2699,12 +2700,12 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         # NOTE: Important that this comes after registerize/registerizeHarder
         if shared.Settings.ELIMINATE_DUPLICATE_FUNCTIONS and shared.Settings.OPT_LEVEL >= 2:
           optimizer.flush()
-          shared.Building.eliminate_duplicate_funcs(final)
+          building.eliminate_duplicate_funcs(final)
           save_intermediate('dfe')
 
       if shared.Settings.EVAL_CTORS and options.memory_init_file and not use_source_map(options) and not shared.Settings.WASM:
         optimizer.flush()
-        shared.Building.eval_ctors(final, memfile)
+        building.eval_ctors(final, memfile)
         save_intermediate('eval-ctors')
 
       if options.js_opts:
@@ -2726,8 +2727,8 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         logger.debug('running closure')
         # no need to add this to js_transform_tempfiles, because closure and
         # debug_level > 0 are never simultaneously true
-        final = shared.Building.closure_compiler(final, pretty=shared.Settings.DEBUG_LEVEL >= 1,
-                                                 extra_closure_args=options.closure_args)
+        final = building.closure_compiler(final, pretty=shared.Settings.DEBUG_LEVEL >= 1,
+                                          extra_closure_args=options.closure_args)
         save_intermediate('closure')
 
     log_time('js opts')
@@ -3130,7 +3131,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
   logger.debug('using binaryen')
   if use_source_map(options) and not shared.Settings.SOURCE_MAP_BASE:
     logger.warning("Wasm source map won't be usable in a browser without --source-map-base")
-  binaryen_bin = shared.Building.get_binaryen_bin()
+  binaryen_bin = building.get_binaryen_bin()
   # whether we need to emit -g (function name debug info) in the final wasm
   debug_info = shared.Settings.DEBUG_LEVEL >= 2 or options.profiling_funcs
   # whether we need to emit -g in the intermediate binaryen invocations (but not necessarily at the very end).
@@ -3141,7 +3142,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
   if not shared.Settings.WASM_BACKEND:
     if DEBUG:
       # save the asm.js input
-      shared.Building.save_intermediate(asm_target, 'asmjs.js')
+      building.save_intermediate(asm_target, 'asmjs.js')
     cmd = [os.path.join(binaryen_bin, 'asm2wasm'), asm_target, '--total-memory=' + str(shared.Settings.INITIAL_MEMORY)]
     if shared.Settings.BINARYEN_TRAP_MODE not in ('js', 'clamp', 'allow'):
       exit_with_error('invalid BINARYEN_TRAP_MODE value: ' + shared.Settings.BINARYEN_TRAP_MODE + ' (should be js/clamp/allow)')
@@ -3150,7 +3151,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       cmd += ['--ignore-implicit-traps']
     # pass optimization level to asm2wasm (if not optimizing, or which passes we should run was overridden, do not optimize)
     if shared.Settings.OPT_LEVEL > 0:
-      cmd.append(shared.Building.opt_level_to_str(shared.Settings.OPT_LEVEL, shared.Settings.SHRINK_LEVEL))
+      cmd.append(building.opt_level_to_str(shared.Settings.OPT_LEVEL, shared.Settings.SHRINK_LEVEL))
     # import mem init file if it exists, and if we will not be using asm.js as a binaryen method (as it needs the mem init file, of course)
     mem_file_exists = options.memory_init_file and os.path.exists(memfile)
     import_mem_init = mem_file_exists and shared.Settings.MEM_INIT_IN_WASM
@@ -3170,7 +3171,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       cmd += ['--mem-max=' + str(shared.Settings.MAXIMUM_MEMORY)]
     if shared.Settings.LEGALIZE_JS_FFI != 1:
       cmd += ['--no-legalize-javascript-ffi']
-    if shared.Building.is_wasm_only():
+    if building.is_wasm_only():
       cmd += ['--wasm-only'] # this asm.js is code not intended to run as asm.js, it is only ever going to be wasm, and can contain special fastcomp-wasm support
     if shared.Settings.USE_PTHREADS:
       cmd += ['--enable-threads']
@@ -3186,7 +3187,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       cmd += ['-o', wasm_binary_target]
     else:
       cmd += ['-o', wasm_text_target, '-S']
-    cmd += shared.Building.get_binaryen_feature_flags()
+    cmd += building.get_binaryen_feature_flags()
     logger.debug('asm2wasm (asm.js => WebAssembly): ' + ' '.join(cmd))
     TimeLogger.update()
     shared.check_call(cmd)
@@ -3223,12 +3224,12 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       options.binaryen_passes += ['--pass-arg=emscripten-sbrk-ptr@%d' % shared.Settings.DYNAMICTOP_PTR]
       if shared.Settings.STANDALONE_WASM:
         options.binaryen_passes += ['--pass-arg=emscripten-sbrk-val@%d' % shared.Settings.DYNAMIC_BASE]
-    shared.Building.save_intermediate(wasm_binary_target, 'pre-byn.wasm')
+    building.save_intermediate(wasm_binary_target, 'pre-byn.wasm')
     args = options.binaryen_passes
-    shared.Building.run_wasm_opt(wasm_binary_target,
-                                 wasm_binary_target,
-                                 args=args,
-                                 debug=intermediate_debug_info)
+    building.run_wasm_opt(wasm_binary_target,
+                          wasm_binary_target,
+                          args=args,
+                          debug=intermediate_debug_info)
 
   if shared.Settings.BINARYEN_SCRIPTS:
     binaryen_scripts = os.path.join(shared.BINARYEN_ROOT, 'scripts')
@@ -3242,8 +3243,8 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
       logger.debug('running binaryen script: ' + script)
       shared.check_call([shared.PYTHON, os.path.join(binaryen_scripts, script), final, wasm_text_target], env=script_env)
   if shared.Settings.EVAL_CTORS:
-    shared.Building.save_intermediate(wasm_binary_target, 'pre-ctors.wasm')
-    shared.Building.eval_ctors(final, wasm_binary_target, binaryen_bin, debug_info=intermediate_debug_info)
+    building.save_intermediate(wasm_binary_target, 'pre-ctors.wasm')
+    building.eval_ctors(final, wasm_binary_target, binaryen_bin, debug_info=intermediate_debug_info)
 
   # after generating the wasm, do some final operations
   if shared.Settings.SIDE_MODULE and not shared.Settings.WASM_BACKEND:
@@ -3260,32 +3261,32 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
 
   # pthreads memory growth requires some additional JS fixups
   if shared.Settings.USE_PTHREADS and shared.Settings.ALLOW_MEMORY_GROWTH:
-    final = shared.Building.apply_wasm_memory_growth(final)
+    final = building.apply_wasm_memory_growth(final)
 
   # >=2GB heap support requires pointers in JS to be unsigned. rather than
   # require all pointers to be unsigned by default, which increases code size
   # a little, keep them signed, and just unsign them here if we need that.
   if shared.Settings.CAN_ADDRESS_2GB:
-    final = shared.Building.use_unsigned_pointers_in_js(final)
+    final = building.use_unsigned_pointers_in_js(final)
 
   if shared.Settings.USE_ASAN:
-    final = shared.Building.instrument_js_for_asan(final)
+    final = building.instrument_js_for_asan(final)
 
   if shared.Settings.OPT_LEVEL >= 2 and shared.Settings.DEBUG_LEVEL <= 2:
     # minify the JS
     optimizer.do_minify() # calculate how to minify
     save_intermediate_with_wasm('preclean', wasm_binary_target)
-    final = shared.Building.minify_wasm_js(js_file=final,
-                                           wasm_file=wasm_binary_target,
-                                           expensive_optimizations=will_metadce(options),
-                                           minify_whitespace=optimizer.minify_whitespace,
-                                           debug_info=intermediate_debug_info)
+    final = building.minify_wasm_js(js_file=final,
+                                    wasm_file=wasm_binary_target,
+                                    expensive_optimizations=will_metadce(options),
+                                    minify_whitespace=optimizer.minify_whitespace,
+                                    debug_info=intermediate_debug_info)
     save_intermediate_with_wasm('postclean', wasm_binary_target)
 
   if shared.Settings.ASYNCIFY_LAZY_LOAD_CODE:
     if not shared.Settings.ASYNCIFY:
       exit_with_error('ASYNCIFY_LAZY_LOAD_CODE requires ASYNCIFY')
-    shared.Building.asyncify_lazy_load_code(wasm_binary_target, debug=intermediate_debug_info)
+    building.asyncify_lazy_load_code(wasm_binary_target, debug=intermediate_debug_info)
 
   def preprocess_wasm2js_script():
     wasm2js = read_and_preprocess(shared.path_from_root('src', 'wasm2js.js'))
@@ -3299,8 +3300,8 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
     return wasm2js
 
   def run_closure_compiler(final):
-    final = shared.Building.closure_compiler(final, pretty=not optimizer.minify_whitespace,
-                                             extra_closure_args=options.closure_args)
+    final = building.closure_compiler(final, pretty=not optimizer.minify_whitespace,
+                                      extra_closure_args=options.closure_args)
     save_intermediate_with_wasm('closure', wasm_binary_target)
     return final
 
@@ -3316,13 +3317,13 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
     else:
       wasm2js_template = final
 
-    wasm2js = shared.Building.wasm2js(wasm2js_template,
-                                      wasm_binary_target,
-                                      opt_level=shared.Settings.OPT_LEVEL,
-                                      minify_whitespace=optimizer.minify_whitespace,
-                                      use_closure_compiler=options.use_closure_compiler,
-                                      debug_info=debug_info,
-                                      symbols_file=symbols_file)
+    wasm2js = building.wasm2js(wasm2js_template,
+                               wasm_binary_target,
+                               opt_level=shared.Settings.OPT_LEVEL,
+                               minify_whitespace=optimizer.minify_whitespace,
+                               use_closure_compiler=options.use_closure_compiler,
+                               debug_info=debug_info,
+                               symbols_file=symbols_file)
     if shared.Settings.WASM == 2:
       shutil.copyfile(wasm2js, wasm2js_template)
       shared.try_delete(wasm2js)
@@ -3339,7 +3340,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
   # note that if we aren't emitting a binary (like in wasm2js) then we don't
   # have anything to do here.
   if options.emit_symbol_map and os.path.exists(wasm_binary_target):
-    shared.Building.handle_final_wasm_symbols(wasm_file=wasm_binary_target, symbols_file=symbols_file, debug_info=debug_info)
+    building.handle_final_wasm_symbols(wasm_file=wasm_binary_target, symbols_file=symbols_file, debug_info=debug_info)
     save_intermediate_with_wasm('symbolmap', wasm_binary_target)
 
   if shared.Settings.DEBUG_LEVEL >= 3 and shared.Settings.SEPARATE_DWARF and os.path.exists(wasm_binary_target):
@@ -3347,7 +3348,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
     dwarf_target = shared.Settings.SEPARATE_DWARF
     if dwarf_target is True:
       dwarf_target = wasm_binary_target + '.debug.wasm'
-    shared.Building.emit_debug_on_side(wasm_binary_target, dwarf_target)
+    building.emit_debug_on_side(wasm_binary_target, dwarf_target)
 
   # replace placeholder strings with correct subresource locations
   if shared.Settings.SINGLE_FILE:
@@ -3769,7 +3770,7 @@ def process_libraries(libs, lib_dirs, temp_files):
       if found:
         break
     if not found:
-      jslibs = shared.Building.path_to_system_js_libraries(lib)
+      jslibs = building.path_to_system_js_libraries(lib)
       if jslibs:
         libraries += [(i, jslib) for jslib in jslibs]
         consumed.append(i)
