+++ tests/test_other.py
@@ -25,24 +25,25 @@ import time
 import tempfile
 import unittest
 import uuid
+from subprocess import PIPE, STDOUT
 
 if __name__ == '__main__':
   raise Exception('do not run this file directly; do something like: tests/runner.py other')
 
-from tools.shared import Building, PIPE, run_js, run_process, STDOUT, try_delete, listify
+from tools.shared import run_js, run_process, try_delete, listify
 from tools.shared import EMCC, EMXX, EMAR, EMRANLIB, PYTHON, FILE_PACKAGER, WINDOWS, LLVM_ROOT, EM_BUILD_VERBOSE
 from tools.shared import CLANG_CC, CLANG_CXX, LLVM_AR, LLVM_DWARFDUMP
 from tools.shared import NODE_JS, SPIDERMONKEY_ENGINE, JS_ENGINES, WASM_ENGINES, V8_ENGINE
 from runner import RunnerCore, path_from_root, no_wasm_backend, no_fastcomp, is_slow_test, ensure_dir
 from runner import needs_dlfcn, env_modify, no_windows, chdir, with_env_modify, create_test_file, parameterized
-from tools import jsrun, shared
+from tools import jsrun, shared, building
 import clang_native
 import tools.line_endings
 import tools.js_optimizer
 import tools.tempfiles
 import tools.duplicate_function_eliminator
 
-scons_path = Building.which('scons')
+scons_path = shared.which('scons')
 emmake = shared.bat_suffix(path_from_root('emmake'))
 emcmake = shared.bat_suffix(path_from_root('emcmake'))
 emconfigure = shared.bat_suffix(path_from_root('emconfigure'))
@@ -97,7 +98,7 @@ def is_python3_version_supported():
   support circular dependencies."""
   try:
     print('is_python3_version_supported')
-    python3 = Building.which('python3')
+    python3 = shared.which('python3')
     print('  python3 =', python3)
     output = run_process([python3, '--version'], stdout=PIPE).stdout
     print('  output =', output, output.split())
@@ -136,7 +137,7 @@ def get_fastcomp_src_dir():
 
 
 def parse_wasm(filename):
-  wat = run_process([os.path.join(Building.get_binaryen_bin(), 'wasm-dis'), filename], stdout=PIPE).stdout
+  wat = run_process([os.path.join(building.get_binaryen_bin(), 'wasm-dis'), filename], stdout=PIPE).stdout
   imports = []
   exports = []
   funcs = []
@@ -294,7 +295,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
       if args[-1] == '/dev/null':
         print('(no output)')
         continue
-      syms = Building.llvm_nm(target)
+      syms = building.llvm_nm(target)
       self.assertIn('main', syms.defs)
       if self.is_wasm_backend():
         # wasm backend will also have '__original_main' or such
@@ -506,7 +507,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
     self.assertContained('warning: Assuming object file output in the absence of `-c`', err)
 
     # Should be two symbols (and in the wasm backend, also __original_main)
-    syms = Building.llvm_nm('combined.o')
+    syms = building.llvm_nm('combined.o')
     self.assertIn('main', syms.defs)
     if self.is_wasm_backend():
       self.assertEqual(len(syms.defs), 3)
@@ -566,7 +567,7 @@ f.close()
   def test_emcc_cflags(self):
     output = run_process([EMCC, '--cflags'], stdout=PIPE)
     flags = output.stdout.strip()
-    self.assertContained(' '.join(Building.doublequote_spaces(shared.emsdk_cflags())), flags)
+    self.assertContained(' '.join(building.doublequote_spaces(shared.emsdk_cflags())), flags)
     # check they work
     cmd = [CLANG_CXX, path_from_root('tests', 'hello_world.cpp')] + shlex.split(flags.replace('\\', '\\\\')) + ['-c', '-emit-llvm', '-o', 'a.bc']
     run_process(cmd)
@@ -622,7 +623,7 @@ f.close()
     for generator in generators:
       conf = configurations[generator]
 
-      if not Building.which(conf['build'][0]):
+      if not shared.which(conf['build'][0]):
         # Use simple test if applicable
         print('Skipping %s test for CMake support; build tool found found: %s.' % (generator, conf['build'][0]))
         continue
@@ -656,7 +657,7 @@ f.close()
           print('Failed command: ' + ' '.join(cmd))
           print('Result:\n' + ret.stdout)
           self.fail('make failed!')
-        self.assertExists(tempdirname + '/' + output_file, 'Building a cmake-generated Makefile failed to produce an output file %s!' % tempdirname + '/' + output_file)
+        self.assertExists(tempdirname + '/' + output_file, 'building a cmake-generated Makefile failed to produce an output file %s!' % tempdirname + '/' + output_file)
 
         # Run through node, if CMake produced a .js file.
         if output_file.endswith('.js'):
@@ -710,16 +711,16 @@ f.close()
     self.clear()
     run_process([emcmake, 'cmake', path_from_root('tests', 'cmake', 'static_lib')])
     run_process(['cmake', '--build', '.'])
-    self.assertTrue(Building.is_ar('libstatic_lib.a'))
+    self.assertTrue(building.is_ar('libstatic_lib.a'))
     run_process([EMAR, 'x', 'libstatic_lib.a'])
     found = False # hashing makes the object name random
     for x in os.listdir('.'):
       if x.endswith('.o'):
         found = True
         if self.is_wasm_backend():
-          assert Building.is_wasm(x)
+          assert building.is_wasm(x)
         else:
-          assert Building.is_bitcode(x)
+          assert building.is_bitcode(x)
     assert found
 
     # Test that passing the -DEMSCRIPTEN_GENERATE_BITCODE_STATIC_LIBRARIES=ON
@@ -729,10 +730,10 @@ f.close()
     run_process([emcmake, 'cmake', '-DEMSCRIPTEN_GENERATE_BITCODE_STATIC_LIBRARIES=ON', path_from_root('tests', 'cmake', 'static_lib')])
     run_process(['cmake', '--build', '.'])
     if self.is_wasm_backend():
-      assert Building.is_wasm('libstatic_lib.bc')
+      assert building.is_wasm('libstatic_lib.bc')
     else:
-      assert Building.is_bitcode('libstatic_lib.bc')
-    assert not Building.is_ar('libstatic_lib.bc')
+      assert building.is_bitcode('libstatic_lib.bc')
+    assert not building.is_ar('libstatic_lib.bc')
 
     # Test that one is able to fake custom suffixes for static libraries.
     # (sometimes projects want to emulate stuff, and do weird things like files
@@ -740,7 +741,7 @@ f.close()
     self.clear()
     run_process([emcmake, 'cmake', '-DSET_FAKE_SUFFIX_IN_PROJECT=1', path_from_root('tests', 'cmake', 'static_lib')])
     run_process(['cmake', '--build', '.'])
-    assert Building.is_ar('myprefix_static_lib.somecustomsuffix')
+    assert building.is_ar('myprefix_static_lib.somecustomsuffix')
 
   # Tests that the CMake variable EMSCRIPTEN_VERSION is properly provided to user CMake scripts
   def test_cmake_emscripten_version(self):
@@ -1044,12 +1045,12 @@ int main() {
       }
     ''')
 
-    Building.emcc(lib_name, output_filename='libA.so')
+    building.emcc(lib_name, output_filename='libA.so')
 
-    Building.emcc(a2_name, ['-L.', '-lA'])
-    Building.emcc(b2_name, ['-L.', '-lA'])
+    building.emcc(a2_name, ['-L.', '-lA'])
+    building.emcc(b2_name, ['-L.', '-lA'])
 
-    Building.emcc(main_name, ['-L.', '-lA', a2_name + '.o', b2_name + '.o'], output_filename='a.out.js')
+    building.emcc(main_name, ['-L.', '-lA', a2_name + '.o', b2_name + '.o'], output_filename='a.out.js')
 
     self.assertContained('result: 1', run_js('a.out.js'))
 
@@ -1076,14 +1077,14 @@ int main() {
     main_name = 'main.c'
     create_test_file(main_name, main)
 
-    Building.emcc(a_name) # a.c.o
-    Building.emcc(b_name) # b.c.o
-    Building.emcc(c_name) # c.c.o
+    building.emcc(a_name) # a.c.o
+    building.emcc(b_name) # b.c.o
+    building.emcc(c_name) # c.c.o
     lib_name = 'libLIB.a'
-    Building.emar('cr', lib_name, [a_name + '.o', b_name + '.o']) # libLIB.a with a and b
+    building.emar('cr', lib_name, [a_name + '.o', b_name + '.o']) # libLIB.a with a and b
 
     # a is in the lib AND in an .o, so should be ignored in the lib. We do still need b from the lib though
-    Building.emcc(main_name, [a_name + '.o', c_name + '.o', '-L.', '-lLIB'], output_filename='a.out.js')
+    building.emcc(main_name, [a_name + '.o', c_name + '.o', '-L.', '-lLIB'], output_filename='a.out.js')
 
     self.assertContained('result: 62', run_js('a.out.js'))
 
@@ -1102,9 +1103,9 @@ int main() {
       }
     ''')
 
-    Building.emcc(lib_src_name) # lib.c.o
+    building.emcc(lib_src_name) # lib.c.o
     lib_name = 'libLIB.a'
-    Building.emar('cr', lib_name, [lib_src_name + '.o']) # libLIB.a with lib.c.o
+    building.emar('cr', lib_name, [lib_src_name + '.o']) # libLIB.a with lib.c.o
 
     def test(lib_args, err_expected):
       print(err_expected)
@@ -1197,11 +1198,11 @@ int f() {
       }
     ''')
 
-    Building.emcc('a.c') # a.c.o
-    Building.emcc('b.c') # b.c.o
-    Building.emcc('c.c')
-    Building.emar('cr', 'libA.a', ['a.c.o', 'c.c.o'])
-    Building.emar('cr', 'libB.a', ['b.c.o', 'c.c.o'])
+    building.emcc('a.c') # a.c.o
+    building.emcc('b.c') # b.c.o
+    building.emcc('c.c')
+    building.emar('cr', 'libA.a', ['a.c.o', 'c.c.o'])
+    building.emar('cr', 'libB.a', ['b.c.o', 'c.c.o'])
 
     args = ['main.c', '-o', 'a.out.js']
     libs_list = ['libA.a', 'libB.a']
@@ -1244,9 +1245,9 @@ int f() {
     main_name = 'main.c'
     create_test_file(main_name, main)
 
-    Building.emcc(lib_name, output_filename='libA.so')
+    building.emcc(lib_name, output_filename='libA.so')
 
-    Building.emcc(main_name, ['libA.so', 'libA.so'], output_filename='a.out.js')
+    building.emcc(main_name, ['libA.so', 'libA.so'], output_filename='a.out.js')
 
     self.assertContained('result: 1', run_js('a.out.js'))
 
@@ -1278,7 +1279,7 @@ int f() {
       };
     ''')
 
-    Building.emcc('lib.c', ['-s', 'EXPORT_ALL', '-s', 'LINKABLE', '--pre-js', 'main.js'], output_filename='a.out.js')
+    building.emcc('lib.c', ['-s', 'EXPORT_ALL', '-s', 'LINKABLE', '--pre-js', 'main.js'], output_filename='a.out.js')
     self.assertContained('libf1\nlibf2\n', run_js('a.out.js'))
 
   def test_export_all_and_exported_functions(self):
@@ -1301,11 +1302,11 @@ int f() {
     ''')
 
     # __get_daylight should not be linked by default, even with EXPORT_ALL
-    Building.emcc('lib.c', ['-s', 'EXPORT_ALL', '--pre-js', 'main.js'], output_filename='a.out.js')
+    building.emcc('lib.c', ['-s', 'EXPORT_ALL', '--pre-js', 'main.js'], output_filename='a.out.js')
     err = run_js('a.out.js', stderr=PIPE, full_output=True, assert_returncode=None)
     self.assertContained('__get_daylight is not defined', err)
 
-    Building.emcc('lib.c', ['-s', "EXPORTED_FUNCTIONS=['__get_daylight']", '-s', 'EXPORT_ALL', '--pre-js', 'main.js'], output_filename='a.out.js')
+    building.emcc('lib.c', ['-s', "EXPORTED_FUNCTIONS=['__get_daylight']", '-s', 'EXPORT_ALL', '--pre-js', 'main.js'], output_filename='a.out.js')
     self.assertContained('libfunc\n', run_js('a.out.js'))
 
   def test_stdin(self):
@@ -1325,16 +1326,16 @@ int f() {
         try_delete('out.txt')
         jscommand = make_js_command(os.path.normpath(exe), engine)
         if WINDOWS:
-          os.system('type "in.txt" | {} >out.txt'.format(' '.join(Building.doublequote_spaces(jscommand))))
+          os.system('type "in.txt" | {} >out.txt'.format(' '.join(building.doublequote_spaces(jscommand))))
         else: # posix
-          os.system('cat in.txt | {} > out.txt'.format(' '.join(Building.doublequote_spaces(jscommand))))
+          os.system('cat in.txt | {} > out.txt'.format(' '.join(building.doublequote_spaces(jscommand))))
         self.assertContained('abcdef\nghijkl\neof', open('out.txt').read())
 
-    Building.emcc(path_from_root('tests', 'module', 'test_stdin.c'), output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'module', 'test_stdin.c'), output_filename='a.out.js')
     create_test_file('in.txt', 'abcdef\nghijkl')
     exe = 'a.out.js'
     run_test()
-    Building.emcc(path_from_root('tests', 'module', 'test_stdin.c'),
+    building.emcc(path_from_root('tests', 'module', 'test_stdin.c'),
                   ['-O2', '--closure', '1'],
                   output_filename='a.out.js')
     run_test()
@@ -1359,7 +1360,7 @@ int f() {
       }
     ''')
     create_test_file('my_test.input', 'abc')
-    Building.emcc('main.cpp', ['--embed-file', 'my_test.input'], output_filename='a.out.js')
+    building.emcc('main.cpp', ['--embed-file', 'my_test.input'], output_filename='a.out.js')
     self.assertContained('zyx', run_process(JS_ENGINES[0] + ['a.out.js'], stdout=PIPE, stderr=PIPE).stdout)
 
   def test_abspaths(self):
@@ -1761,29 +1762,29 @@ int f() {
     self.assertContained('1234, 1234, 4321\n', run_js('a.out.js'))
 
   def test_sdl2_mixer(self):
-    Building.emcc(path_from_root('tests', 'sdl2_mixer.c'), ['-s', 'USE_SDL_MIXER=2'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'sdl2_mixer.c'), ['-s', 'USE_SDL_MIXER=2'], output_filename='a.out.js')
 
   def test_libpng(self):
     shutil.copyfile(path_from_root('tests', 'pngtest.png'), 'pngtest.png')
-    Building.emcc(path_from_root('tests', 'pngtest.c'), ['--embed-file', 'pngtest.png', '-s', 'USE_ZLIB=1', '-s', 'USE_LIBPNG=1'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'pngtest.c'), ['--embed-file', 'pngtest.png', '-s', 'USE_ZLIB=1', '-s', 'USE_LIBPNG=1'], output_filename='a.out.js')
     self.assertContained('TESTS PASSED', run_process(JS_ENGINES[0] + ['a.out.js'], stdout=PIPE, stderr=PIPE).stdout)
 
   def test_libjpeg(self):
     shutil.copyfile(path_from_root('tests', 'screenshot.jpg'), 'screenshot.jpg')
-    Building.emcc(path_from_root('tests', 'jpeg_test.c'), ['--embed-file', 'screenshot.jpg', '-s', 'USE_LIBJPEG=1'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'jpeg_test.c'), ['--embed-file', 'screenshot.jpg', '-s', 'USE_LIBJPEG=1'], output_filename='a.out.js')
     self.assertContained('Image is 600 by 450 with 3 components', run_js('a.out.js', args=['screenshot.jpg'], stdout=PIPE, stderr=PIPE))
 
   def test_bullet(self):
-    Building.emcc(path_from_root('tests', 'bullet_hello_world.cpp'), ['-s', 'USE_BULLET=1'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'bullet_hello_world.cpp'), ['-s', 'USE_BULLET=1'], output_filename='a.out.js')
     self.assertContained('BULLET RUNNING', run_process(JS_ENGINES[0] + ['a.out.js'], stdout=PIPE, stderr=PIPE).stdout)
 
   def test_vorbis(self):
     # This will also test if ogg compiles, because vorbis depends on ogg
-    Building.emcc(path_from_root('tests', 'vorbis_test.c'), ['-s', 'USE_VORBIS=1'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'vorbis_test.c'), ['-s', 'USE_VORBIS=1'], output_filename='a.out.js')
     self.assertContained('ALL OK', run_process(JS_ENGINES[0] + ['a.out.js'], stdout=PIPE, stderr=PIPE).stdout)
 
   def test_bzip2(self):
-    Building.emcc(path_from_root('tests', 'bzip2_test.c'), ['-s', 'USE_BZIP2=1'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'bzip2_test.c'), ['-s', 'USE_BZIP2=1'], output_filename='a.out.js')
     self.assertContained("usage: unzcrash filename", run_process(JS_ENGINES[0] + ['a.out.js'], stdout=PIPE, stderr=PIPE).stdout)
 
   def test_freetype(self):
@@ -1791,7 +1792,7 @@ int f() {
     # <emscripten_root>/tests/freetype to the compilation folder
     shutil.copy2(path_from_root('tests/freetype', 'LiberationSansBold.ttf'), os.getcwd())
     # build test program with the font file embed in it
-    Building.emcc(path_from_root('tests', 'freetype_test.c'), ['-s', 'USE_FREETYPE=1', '--embed-file', 'LiberationSansBold.ttf'], output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'freetype_test.c'), ['-s', 'USE_FREETYPE=1', '--embed-file', 'LiberationSansBold.ttf'], output_filename='a.out.js')
     # the test program will print an ascii representation of a bitmap where the
     # 'w' character has been rendered using the Liberation Sans Bold font
     expectedOutput = '                \n' + \
@@ -2054,7 +2055,7 @@ int f() {
     assert 'hello, world!' in run_js('a.out.js')
     self.assertExists('my_bitcode.bc')
     try_delete('a.out.js')
-    Building.llvm_dis('my_bitcode.bc', 'my_ll.ll')
+    building.llvm_dis('my_bitcode.bc', 'my_ll.ll')
     run_process([EMCC, 'my_ll.ll', '-nostdlib', '-o', 'two.js'])
     assert 'hello, world!' in run_js('two.js')
 
@@ -3756,7 +3757,7 @@ int main()
 
   @no_wasm_backend('fastcomp specific')
   def test_llvm_lit(self):
-    grep_path = Building.which('grep')
+    grep_path = shared.which('grep')
     if not grep_path:
       self.skipTest('This test needs the "grep" tool in PATH. If you are using emsdk on Windows, you can obtain it via installing and activating the gnu package.')
     llvm_src = get_fastcomp_src_dir()
@@ -3972,7 +3973,7 @@ EM_ASM({ _middle() });
           out = run_js('a.out.js', stderr=PIPE, full_output=True, assert_returncode=None)
           self.assertContained(stack_trace_reference, out)
           # make sure there are no symbols in the wasm itself
-          wat = run_process([os.path.join(Building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
+          wat = run_process([os.path.join(building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
           for func_start in ('(func $middle', '(func $_middle'):
             self.assertNotContained(func_start, wat)
         # check we don't keep unnecessary debug info with wasm2js when emitting
@@ -4581,7 +4582,7 @@ int main() {
     self.assertContained('target triple = "', bitcode)
 
     run_process([EMCC, path_from_root('tests', 'hello_world.c'), '-c', '-emit-llvm'])
-    self.assertTrue(Building.is_bitcode('hello_world.bc'))
+    self.assertTrue(building.is_bitcode('hello_world.bc'))
 
   def test_dashE(self):
     create_test_file('src.cpp', r'''#include <emscripten.h>
@@ -4866,7 +4867,7 @@ int main()
     self.assertContained('ok!', run_js('a.out.js'))
 
   def test_strptime_symmetry(self):
-    Building.emcc(path_from_root('tests', 'strptime_symmetry.cpp'), output_filename='a.out.js')
+    building.emcc(path_from_root('tests', 'strptime_symmetry.cpp'), output_filename='a.out.js')
     self.assertContained('TEST PASSED', run_js('a.out.js'))
 
   def test_truncate_from_0(self):
@@ -5516,7 +5517,7 @@ int main(void) {
 
   def test_require(self):
     inname = path_from_root('tests', 'hello_world.c')
-    Building.emcc(inname, args=['-s', 'ASSERTIONS=0'], output_filename='a.out.js')
+    building.emcc(inname, args=['-s', 'ASSERTIONS=0'], output_filename='a.out.js')
     output = run_process(NODE_JS + ['-e', 'require("./a.out.js")'], stdout=PIPE, stderr=PIPE)
     assert output.stdout == 'hello, world!\n' and output.stderr == '', 'expected no output, got\n===\nSTDOUT\n%s\n===\nSTDERR\n%s\n===\n' % (output.stdout, output.stderr)
 
@@ -6858,7 +6859,7 @@ high = 1234
       if python == 'python3':
         has = is_python3_version_supported()
       else:
-        has = Building.which(python) is not None
+        has = shared.which(python) is not None
       print(python, has)
       if has:
         print('  checking emcc.py...')
@@ -6949,10 +6950,10 @@ int main() {
       }
     ''')
 
-    Building.emcc(os.path.join(with_space, 'main.cpp'), ['-g'])
+    building.emcc(os.path.join(with_space, 'main.cpp'), ['-g'])
 
     with chdir(with_space):
-      link_args = Building.link(['main.cpp.o'], 'all.bc', just_calculate=True)
+      link_args = building.link(['main.cpp.o'], 'all.bc', just_calculate=True)
 
     time.sleep(0.2) # Wait for Windows FS to release access to the directory
     shutil.rmtree(with_space)
@@ -7425,11 +7426,11 @@ int main() {
     run_process(cmd)
     # If foo() had been wrongly inlined above, internalizing foo and running
     # global DCE makes foo DCE'd
-    Building.llvm_opt('test.bc', ['-internalize', '-internalize-public-api-list=main', '-globaldce'], 'test2.bc')
+    building.llvm_opt('test.bc', ['-internalize', '-internalize-public-api-list=main', '-globaldce'], 'test2.bc')
 
     # To this test to be successful, foo() shouldn't have been inlined above and
     # foo() should be in the function list
-    syms = Building.llvm_nm('test2.bc', include_internal=True)
+    syms = building.llvm_nm('test2.bc', include_internal=True)
     assert 'foo' in syms.defs, 'foo() should not be inlined'
 
   @no_wasm_backend('--separate-asm')
@@ -7483,7 +7484,7 @@ int main() {
         assert expect_js_opts == ('applying js optimization passes:' in err), err
         if not self.is_wasm_backend():
           assert expect_only_wasm == ('-emscripten-only-wasm' in err and '--wasm-only' in err), err # check both flag to fastcomp and to asm2wasm
-        wat = run_process([os.path.join(Building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
+        wat = run_process([os.path.join(building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
         # i64s
         i64s = wat.count('(i64.')
         print('    seen i64s:', i64s)
@@ -7601,7 +7602,7 @@ int main() {
       cmd = [EMCC, path_from_root('tests', 'hello_world.c'), '-s', 'WASM=1', '-O2'] + args
       print(' '.join(cmd))
       run_process(cmd)
-      wat = run_process([os.path.join(Building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
+      wat = run_process([os.path.join(building.get_binaryen_bin(), 'wasm-dis'), 'a.out.wasm'], stdout=PIPE).stdout
       for line in wat:
         if '(import "env" "memory" (memory ' in line:
           parts = line.strip().replace('(', '').replace(')', '').split(' ')
@@ -7978,7 +7979,7 @@ int main() {
 
   def test_legalize_js_ffi(self):
     # test disabling of JS FFI legalization
-    wasm_dis = os.path.join(Building.get_binaryen_bin(), 'wasm-dis')
+    wasm_dis = os.path.join(building.get_binaryen_bin(), 'wasm-dis')
     for (args, js_ffi) in [
         (['-s', 'LEGALIZE_JS_FFI=1', '-s', 'SIDE_MODULE=1', '-O1', '-s', 'EXPORT_ALL=1'], True),
         (['-s', 'LEGALIZE_JS_FFI=0', '-s', 'SIDE_MODULE=1', '-O1', '-s', 'EXPORT_ALL=1'], False),
@@ -8031,7 +8032,7 @@ int main() {
     # test minimal JS FFI legalization for invoke and dyncalls
     if self.is_wasm_backend():
       self.skipTest('not testing legalize with main module and wasm backend')
-    wasm_dis = os.path.join(Building.get_binaryen_bin(), 'wasm-dis')
+    wasm_dis = os.path.join(building.get_binaryen_bin(), 'wasm-dis')
     for (args, js_ffi) in [
         (['-s', 'LEGALIZE_JS_FFI=0', '-s', 'MAIN_MODULE=2', '-O3', '-s', 'DISABLE_EXCEPTION_CATCHING=0'], False),
       ]:
@@ -8110,7 +8111,7 @@ int main() {
         if target.endswith('.wasm'):
           # only wasm requested
           self.assertNotExists('out.js')
-        wat = run_process([os.path.join(Building.get_binaryen_bin(), 'wasm-dis'), 'out.wasm'], stdout=PIPE).stdout
+        wat = run_process([os.path.join(building.get_binaryen_bin(), 'wasm-dis'), 'out.wasm'], stdout=PIPE).stdout
         wat_lines = wat.split('\n')
         exports = [line.strip().split(' ')[1].replace('"', '') for line in wat_lines if "(export " in line]
         imports = [line.strip().split(' ')[2].replace('"', '') for line in wat_lines if "(import " in line]
@@ -8153,13 +8154,13 @@ int main() {
 
       print('wasm in object')
       run_process([EMXX, src] + args + ['-c', '-o', 'hello_obj.o'])
-      self.assertTrue(shared.Building.is_wasm('hello_obj.o'))
-      self.assertFalse(shared.Building.is_bitcode('hello_obj.o'))
+      self.assertTrue(building.is_wasm('hello_obj.o'))
+      self.assertFalse(building.is_bitcode('hello_obj.o'))
 
       print('bitcode in object')
       run_process([EMXX, src] + args + ['-c', '-o', 'hello_bitcode.o', '-flto'])
-      self.assertFalse(shared.Building.is_wasm('hello_bitcode.o'))
-      self.assertTrue(shared.Building.is_bitcode('hello_bitcode.o'))
+      self.assertFalse(building.is_wasm('hello_bitcode.o'))
+      self.assertTrue(building.is_bitcode('hello_bitcode.o'))
 
       print('use bitcode object (LTO)')
       run_process([EMXX, 'hello_bitcode.o'] + args + ['-flto'])
@@ -8193,7 +8194,7 @@ int main() {
       (['-s', 'WASM_OBJECT_FILES=1'], False),
     ]:
       run_process([EMCC, path_from_root('tests', 'hello_world.cpp')] + flags + ['-c', '-o', 'a.o'])
-      seen_bitcode = Building.is_bitcode('a.o')
+      seen_bitcode = building.is_bitcode('a.o')
       self.assertEqual(expect_bitcode, seen_bitcode, 'must emit LTO-capable bitcode when flags indicate so (%s)' % str(flags))
 
   def test_wasm_nope(self):
@@ -8382,7 +8383,7 @@ end
     # sent via response file
     create_test_file("file'1", ' ')
     create_test_file("file'2", ' ')
-    Building.emar('cr', 'libfoo.a', ("file'1", "file'2"))
+    building.emar('cr', 'libfoo.a', ("file'1", "file'2"))
 
   def test_archive_empty(self):
     # This test added because we had an issue with the AUTO_ARCHIVE_INDEXES failing on empty
@@ -8530,7 +8531,7 @@ end
     env['LC_ALL'] = 'C'
     expected = ': supported targets:.* elf'
     for python in [PYTHON, 'python', 'python2', 'python3']:
-      if not Building.which(python):
+      if not shared.which(python):
         continue
       if python == 'python3' and not is_python3_version_supported():
         continue
@@ -8861,28 +8862,28 @@ test_module().then((test_module_instance) => {
 
     with open(fname, 'wb') as f:
       f.write(b'foo')
-    self.assertFalse(Building.is_bitcode(fname))
+    self.assertFalse(building.is_bitcode(fname))
 
     with open(fname, 'wb') as f:
       f.write(b'\xDE\xC0\x17\x0B')
       f.write(16 * b'\x00')
       f.write(b'BC')
-    self.assertTrue(Building.is_bitcode(fname))
+    self.assertTrue(building.is_bitcode(fname))
 
     with open(fname, 'wb') as f:
       f.write(b'BC')
-    self.assertTrue(Building.is_bitcode(fname))
+    self.assertTrue(building.is_bitcode(fname))
 
   def test_is_ar(self):
     fname = 'tmp.a'
 
     with open(fname, 'wb') as f:
       f.write(b'foo')
-    self.assertFalse(Building.is_ar(fname))
+    self.assertFalse(building.is_ar(fname))
 
     with open(fname, 'wb') as f:
       f.write(b'!<arch>\n')
-    self.assertTrue(Building.is_ar(fname))
+    self.assertTrue(building.is_ar(fname))
 
   def test_emcc_parsing(self):
     create_test_file('src.c', r'''
@@ -10197,9 +10198,9 @@ int main() {
 
   def is_object_file(self, filename):
     if self.is_wasm_backend():
-      return shared.Building.is_wasm('-')
+      return building.is_wasm('-')
     else:
-      return shared.Building.is_bitcode('-')
+      return building.is_bitcode('-')
 
   def test_stdout_link(self):
     # linking to stdout `-` doesn't work, we have no way to pass such an output filename
@@ -10308,9 +10309,9 @@ int main() {
     # Verify that bitcode files are accepted as input
     create_test_file('main.c', 'void foo(); int main() { return 0; }')
     run_process([EMCC, '-emit-llvm', '-c', '-o', 'main.bc', 'main.c'])
-    self.assertTrue(shared.Building.is_bitcode('main.bc'))
+    self.assertTrue(building.is_bitcode('main.bc'))
     run_process([EMCC, '-c', '-o', 'main.o', 'main.bc'])
-    self.assertTrue(shared.Building.is_wasm('main.o'))
+    self.assertTrue(building.is_wasm('main.o'))
 
   def test_nostdlib(self):
     # First ensure all the system libs are built
