+++ tools/duplicate_function_eliminator.py
@@ -16,7 +16,7 @@ import traceback
 
 sys.path.insert(1, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from tools import shared
+from tools import shared, building
 from tools.js_optimizer import DEBUG, temp_files, start_funcs_marker, end_funcs_marker, split_funcs, start_asm_marker, end_asm_marker
 from tools.js_optimizer import MIN_CHUNK_SIZE, MAX_CHUNK_SIZE, NUM_CHUNKS_PER_CORE
 
@@ -237,7 +237,7 @@ def run_on_js(filename, gen_hash_info=False):
 
   # if we are making source maps, we want our debug numbering to start from the
   # top of the file, so avoid breaking the JS into chunks
-  cores = shared.Building.get_num_cores()
+  cores = building.get_num_cores()
 
   intended_num_chunks = int(round(cores * NUM_CHUNKS_PER_CORE))
   chunk_size = min(MAX_CHUNK_SIZE, max(MIN_CHUNK_SIZE, total_size / intended_num_chunks))
@@ -273,7 +273,7 @@ def run_on_js(filename, gen_hash_info=False):
       # We can parallelize
       if DEBUG:
         print('splitting up js optimization into %d chunks, using %d cores  (total: %.2f MB)' % (len(chunks), cores, total_size / (1024 * 1024.)), file=sys.stderr)
-      pool = shared.Building.get_multiprocessing_pool()
+      pool = building.get_multiprocessing_pool()
       filenames = pool.map(run_on_chunk, commands, chunksize=1)
     else:
       # We can't parallize, but still break into chunks to avoid uglify/node memory issues
